requirements.txt
Django>=5.2,<5.3
psycopg[binary]>=3.1,<4
gunicorn>=22.0,<23
django-htmx>=1.21,<2
djangorestframework>=3.15,<4
django-filter>=24.0,<25
django-simple-history>=3.7,<4
djangorestframework-simplejwt>=5.3,<6
openpyxl>=3.1,<4
